test: move folder with tests into src/
