+++ src/tests/Header.spec.tsx
@@ -1,6 +1,6 @@
 import { mount } from '@cypress/react'
 import { BrowserRouter } from 'react-router-dom'
-import Header from '../Header'
+import Header from '../components/Header'
 
 let isLogged = true
